pyproject.toml

[tool.pytest.ini_options] 
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s | %(filename)s:%(lineno)d:%(function)s | %(levelname)s | %(message)s"

[tool.poetry]
name = "compare-my-stocks"
version = "1.0.7"
description = "A system for visualizing interesting stocks. Has powerful comparison capabilities and works seamlessly with your jupyter notebook.   Written in QT with matplotlib."
authors = ["eyalk5"]
readme = "README.md"
packages = [{include = "compare_my_stocks", from = "src"}]
include = [  "src/compare_my_stocks/data",
"src/compare_my_stocks/gui/mainwindow.ui",
"src/compare_my_stocks/common/impacketLICENSE"]
exclude = [
    "tests/not_this_dir"
]

[tool.poetry.dependencies]
colorlog = ">=6.7.0,<6.8.0"
django = ">=3.2.9,<3.2.10"
numpy = ">=1.22.0,<1.23.0"
"ruamel.yaml" = "^0.17.31"
pydantic = "^2.3.0"

dacite = "^1.8.1"
python = ">=3.9,<=3.11"

requests = {version=">=2.28.0", optional= true}
pyside6 = {version="^6.5.2", optional= true}
matplotlib = {version="^3.7.2", optional= true}
pandas = {version="1.5.3", optional= true} #the objects were pickled with this version
python-dateutil = {version=">=2.8.2,<2.8.3", optional= true}
pytz = {version=">=2021.3,<2022.0", optional= true}
qtvoila = {version="2.1.0", optional= true}
superqt = {version=">=0.2.5.post1,<0.3.0", optional= true}
mplcursors = {version=">=0.4,<1.0", optional= true}
ib-insync = {version="^0.9.85", optional= true}
pyro5 = {version=">=5.14,<6.0", optional= true}
ibflex = {version=">=0.15", optional= true}
six = {version=">=1.16.0,<1.17.0", optional= true}
psutil = {version=">=5.8.0,<5.9.0", optional= true}
dataconf = {version=">=2.1.3,<2.1.4", optional= true}
toml = {version=">=0.10.2,<0.11.0", optional= true}
pytest = {version="^7.3.0", optional= true}
ipykernel = {version="*", optional= true}
setuptools = {version=">=60.2.0,<60.3.0", optional= true}
nbformat = {version=">=5.7.3,<5.8.0", optional= true}
memoization = {version="0.4.0", optional= true}
notebook = { version = "^6.4.2", optional= true}
json-editor-pyside6 = {version="^1.0.3", optional= true}
qt-collapsible-section-pyside6 = {version="^0.1.0", optional= true}
jupyter-server = { version="^2.7.3", optional= true} # needed really by notebook. just to simply finding.
ipydatagrid = { version= "*",optional = true }
joblib = { version= "^1.3.2",optional = true }
polygon-api-client = {version ="^1.12.4",optional = true} 
multimethod = { version ="*",optional = true}
ordered-set = { version ="*",optional = true}


voila = {version="*", optional= true}
python-composition = {version="*", optional= true}

[tool.poetry.extras]
full = ['requests', 'pyside6', 'matplotlib', 'pandas', 'python-dateutil', 'pytz', 'qtvoila', 'superqt', 'mplcursors', 'ib-insync', 'pyro5', 'ibflex', 'six', 'psutil', 'dataconf', 'toml', 'pytest', 'ipykernel', 'setuptools', 'nbformat', 'memoization', 'notebook', 'json-editor-pyside6', 'qt-collapsible-section-pyside6', 'jupyter-server', 'ipydatagrid', 'joblib', 'polygon-api-client', 'multimethod', 'ordered-set', 'voila', 'python-composition']
mini = ['colorlog','django','numpy','ruamel.yaml']
jupyter = ['ipydatagrid','colorlog','django','numpy','ruamel.yaml','voila','ipykernel','pandas','voila','joblib','matplotlib']
full_no_ib = ['ipydatagrid', 'requests','pyside6','matplotlib','pandas','python','pytz','json-editor-pyside6','qtvoila','superqt','mplcursors','pyro5','qt-collapsible-section-pyside6','six','psutil','dataconf','toml','pytest','ipykernel','setuptools','nbformat', 'colorlog','django','numpy','ruamel.yaml','notebook','joblib','matplotlib','polygon-api-client', 'joblib', 'polygon-api-client', 'multimethod', 'ordered-set', 'voila', 'python-composition']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
[mypy]
mypy_path = "src\\compare_my_stocks"
[tool.pyright]
pythonVersion= "3.9.6"
ignore = ["./venv"]
